refactor norm and file
