+++ source/data_struct/free_all.c
@@ -6,7 +6,7 @@
 /*   By: ebennace <ebennace@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/11 14:39:15 by ebennace          #+#    #+#             */
-/*   Updated: 2022/10/18 18:39:15 by ebennace         ###   ########.fr       */
+/*   Updated: 2022/10/23 18:41:39 by ebennace         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
