--- source/get/get_philo.c
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   get_philo.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ebennace <ebennace@student.42lausanne.c    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/11 15:01:07 by ebennace          #+#    #+#             */
-/*   Updated: 2022/10/20 12:13:11 by ebennace         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../header/philosopher.h"
-
-t_philo *get_first_philo(t_env *env)
-{
-    if (env->first_philo)
-        return (env->first_philo);
-    return (NULL);
-}
