+++ source/is/is_philo.c
@@ -6,7 +6,7 @@
 /*   By: ebennace <ebennace@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/20 11:53:22 by ebennace          #+#    #+#             */
-/*   Updated: 2022/10/20 11:53:31 by ebennace         ###   ########.fr       */
+/*   Updated: 2022/10/23 18:37:41 by ebennace         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
